fix bugs: 1. Can not get channel_layout when read wav files so that swr context can't be init. Now use av_get_default_channel_layout(in_channels) to get channel layout if in_channel_layout == 0. 2. Duration of AAC file is not correct. I write a AACUtil.cpp to calculate aac duration.
Now it can play wav, mp3, aac, ac3, ape, m4a, amr, 3gp and so on. But ape pts is not always usable and it will cause updating progress every 2 seconds.

+++ app/src/main/java/com/zu/ffmpegaudioplayer/SongListAdapter.kt
@@ -9,9 +9,9 @@ import androidx.recyclerview.widget.RecyclerView
 import com.zu.ffmpegaudioplayer.data.AudioFile
 import java.text.SimpleDateFormat
 
-fun formatDuration(duration: Int): String
+fun formatDuration(duration: Long): String
 {
-    val totalSeconds: Int = (duration / 1000)
+    val totalSeconds: Int = (duration / 1000).toInt()
     val totalMinutes = totalSeconds / 60
     val seconds: Int = totalSeconds % 60
     val minutes: Int = totalMinutes % 60
@@ -63,7 +63,7 @@ class SongListAdapter: RecyclerView.Adapter<SongListItem>()
         }
 
         holder.tvName.text = data!![position].fileName
-        holder.tvDuration.text = formatDuration(data!![position].duration!!)
+        holder.tvDuration.text = formatDuration(data!![position].duration!!.toLong())
         holder.itemView.tag = position
 
         holder.itemView.setOnClickListener(internalItemClickerListener)
